pithos: Restore XML input in uploads using hashmaps
In version 0.15 the option to create an object by providing
the object's hashmap in XML format, instead of actual data, was disabled.
This commit restores this operation by substituting
the minidom module (vulnerable to some security implications)
with the lxml module for performing the XML parsing.
It fixes also the respective test case.

Conflicts:
	snf-pithos-app/pithos/api/test/objects.py
	snf-pithos-app/setup.py